+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bank accountlijst overflowed niet meer en is nu scrollable
 - Je kan niet meer cornersellen aan bepaalde NPCs
 - Je kan motors nu wel afleggen als je sleutels niet hebt
+- Betere manier om admin screenshot te kopieren
 
 ## [1.0.9]